Tarefa5

Description

Introdução

Foi nesta tarefa que tivemos a primeora experiência com inteligência artificial dentro do trabalho. Sendo por isso uma das tarefas maios complicadas do trabalho. Apesar de ser díficil é sempre desafiante fazer algo deste género. Após umas largas horas de testes e de escrita de código fomos aprendendo como é que poderíamos fazer acontecer.

Objetivos

O grande objetivo da tarefa passava por fazer com que os fantasmas decidissem a sua próxima jogada tendo em conta todas as variáveis que os rodeiam. Assim sendo, os nossos fantasmas poderiam adotar dois tipos de comportamentos, uma de fugir e outra de andar atrás do pacman, dependendo do modo em que se encontrasse. Bastava então perceber como é que colocariamos isso no código. Decidimos então fazer dois tipos de algoritmos, um deles composto por uma RTree, assim ele poderia tomar a sua decisão conforme a sua distância ao pacman e o seu modo. E muito resumidamente foi assim que conseguimos programar os nossos fantasmas para fazer o que lhes competia.

Discussão e conclusão

Sendo que esta, já como acima referido, foi uma das tarefas mais dificeis de concluir fomos tendo alguns erros, maas conseguimos ultrapassa-los, fazendo com que o fantasma agisse como pretendido

Documentation

jogador3 :: Player

enemy3 :: Player

testMaze :: Maze

testState :: State

Funções principais

data RTree a

Constructors

T a [RTree a]

profundidade

Arguments

:: Int
Uma constante

fillLeaf

Arguments

:: Maze
O Maze

-> RTree (Coords, Orientation)
A RTree inicial

-> RTree (Coords, Orientation)
A RTree com as folhas preenchidas

Preenche as folhas da RTree

fillPath

Arguments

:: Coords
Coords alvo

-> Int
A profundidade

-> Maze
O Maze

-> [Coords]
A lista de todas as coordenadas dos "pais"

-> RTree (Coords, Orientation)
A RTree inicial

-> RTree (Coords, Orientation)
A RTree final

...

areCoordsEqual

Arguments

:: Coords
Umas quaisquer Coords

-> Coords
Outras Coords quaisqueres

-> Bool
O Bool correspondente

Compara duas Coords e vê se elas são iguais

getNextMove

Arguments

:: Int
O que irá decidir o comportamento do bot

-> Maze
O Maze

-> Orientation
A Orientation

-> [Coords]
Coords da casa dos fantasmas

-> Coords
Coords iniciais

-> Coords
Coords alvo

-> Orientation
A Orientation final

Escolhe o próximo movimento do Bot

getSmallestPath

Arguments

:: [[a]]
O conjunto de arrays

-> [a]
Um acumulador

-> [a]
O array mais pequeno

Escolhe o menor array dentro de um conjunto de arrays

getHeadInfo

Arguments

:: RTree (Coords, Orientation)
Uma RTree

-> (Coords, Orientation)
As Coords e Orientation do primeiro elemento

Recolhe a informação sobre o primeiro tuplo da RTree

getDistanceTo

Arguments

:: Coords
Umas Coords

-> RTree (Coords, Orientation)
Uma RTree

-> (Double, RTree (Coords, Orientation))
Um tuplo da distância e da RTree

Descobre a distância entre duas Coords devolvendo a distância e a RTree dada

pathsToNode

Arguments

:: (Coords, Orientation)

-> RTree (Coords, Orientation)
Uma RTree

-> [[(Coords, Orientation)]]
A lista de Coords e Orientation

Retorna o caminho para um determinado node numa RTree

getAllPlayersCoords

Arguments

:: State
Um State

-> [Coords]
A lista das Coords de todos os Players

Descobre as Coords de todos os Players

distance

Arguments

:: Coords
As Coords de um Ghost

-> Coords
As Coords de um Player

-> Double
A distância entre eles

Calcula a distância entre duas Coords

getNeighboursTree

Arguments

:: Maze
Um Maze

-> Coords
Umas Coords

-> [RTree (Coords, Orientation)]
Uma RTree dos vizinhos das Coords anteriores

Consegue a árvore dos vizinhos das Coords anteriores

getNeighbours

Arguments

:: Maze
Um Maze

-> [Coords]
Uma lista de Coords

-> Coords
Umas Coords

-> [(Coords, Orientation)]
Um tuplo com Coords e Orientation

Descobre todos os vizinhos

alturaRTree

Arguments

:: RTree (Coords, Orientation)
Uma RTree

-> Int
A sua altura

Determina a altura da RTree

evaluateMoves :: Int -> Maze -> Coords -> [(Coords, Orientation)] -> (Int, Coords) -> Orientation

getNextPlay

Arguments

:: [Player]
Uma lista de Players

-> Int
O que irá decidir o comportamento do bot

-> Int
O Id do Ghost

-> Maze
Um Maze

-> Orientation
A Orientation

-> Coords
As Coords do Ghost

-> [(Coords, Orientation)]
Uma lista de tuplos de Coords e Orientation

-> (Int, Coords)
O Id e Coords do 'Player

-> [Coords]

-> Play
O Play

Gera a próxima jogada de um Player

getOrientationFromTuple

Arguments

:: Int
O mesmo int que determina a procura no tuple

-> [(Double, Orientation)]
Uma lista de tuplos

-> Orientation
A Orientation obtida do tuplo

Tira a Orientation de um tuplo

getOrientationByID

Arguments

:: [Player]
Uma lista de Players

-> Int

-> Orientation

Tira a Orientation de um Id

getCoordsByID

Arguments

:: [Player]
A lista de PLayers

-> Int
O ID

-> Coords
As Coords correspondentes ao Id

A partir do Id de um Player conseguir as Coords desse mesmo player

getPacModeByID

Arguments

:: [Player]
A lista de PLayers

-> Int
O ID

-> PacMode
As Coords correspondentes ao Id

A partir do ID de um Player conseguir o seu PacMode

getClosestGhost

Arguments

:: [Player]
A lista de Players

-> Coords
As Coords

-> Int
O acumulador do ID

-> Coords
O acumulador das Coords

-> (Int, Coords)
Um tuplo com o ID e as Coords do Ghost

Encontra o Ghost mais próximo das Coords dadas

getClosestPlayer

Arguments

:: [Player]
A lista de Players

-> Coords
As Coords

-> Int
O acumulador do ID

-> Coords
O acumulador das Coords

-> (Int, Coords)
Um tuplo com o ID e as Coords do Pacman

Encontra o Pacman mais próximo das Coords dadas

chaseMode

Arguments

:: State
Um State

-> [Coords]
As Coords da casa dos fantasmas

-> Int
O Id

-> Play
O Play

Neste modo ele vai perseguir o Pacman

scatterMode

Arguments

:: State
Um State

-> [Coords]
As Coords da casa dos fantasmas

-> Int
O Id

-> Play
O Play

Neste modo ele vai fugir do Pacman

ghostPlay

Arguments

:: State
Um State

-> [Play]
A lista de Plays

Onde o ghost efetua as jogadas